+++ Assets/Use Case Samples/Battle Pass/Scripts/CloudCodeManager.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Unity.Services.CloudCode;
 using UnityEngine;